+++ api/v1/nifiparametercontext_types.go
@@ -42,6 +42,8 @@ type NifiParameterContextStatus struct {
 	Version int64 `json:"version"`
 	// the latest update request.
 	LatestUpdateRequest *ParameterContextUpdateRequest `json:"latestUpdateRequest,omitempty"`
+	// the latest update time of the secrets.
+	LastestSecretUpdate *metav1.Time `json:"latestSecretUpdate,omitempty"`
 }
 
 type ParameterContextUpdateRequest struct {